Landing fragment (#12)
* Home Fragment Updated and Created FormatUtils.kt

+++ app/src/main/java/com/tec9ers/thunderstorm/model/onecallapi/Current.kt
@@ -9,10 +9,10 @@ data class Current (
 	val dt : Int,
 
 	@SerializedName("sunrise")
-	val sunrise : Int,
+	val sunrise : Long,
 
 	@SerializedName("sunset")
-	val sunset : Int,
+	val sunset : Long,
 
 	@SerializedName("temp")
 	val temp : Double,
